+++ CHANGELOG.md
@@ -1,5 +1,46 @@
 # Changelog
 
+## v0.9.5
+
+[compare changes](https://github.com/h3js/srvx/compare/v0.9.3...v0.9.5)
+
+### 🚀 Enhancements
+
+- **node, deno, bun:** Validate server port range ([#137](https://github.com/h3js/srvx/pull/137))
+
+### 🩹 Fixes
+
+- Handle non promise and promise-like values for `waitUntil` ([ef71acd](https://github.com/h3js/srvx/commit/ef71acd))
+
+### 💅 Refactors
+
+- Improve gracefulShutdown dx ([3170f23](https://github.com/h3js/srvx/commit/3170f23))
+
+### 📦 Build
+
+- Better tree-shaking cli colors ([0973a59](https://github.com/h3js/srvx/commit/0973a59))
+
+### 🏡 Chore
+
+- **release:** V0.9.4 ([70917bf](https://github.com/h3js/srvx/commit/70917bf))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](https://github.com/pi0))
+- Abeer0 ([@iiio2](https://github.com/iiio2))
+
+## v0.9.4
+
+[compare changes](https://github.com/h3js/srvx/compare/v0.9.3...v0.9.4)
+
+### 🩹 Fixes
+
+- Handle non promise and promise-like values for `waitUntil` ([ef71acd](https://github.com/h3js/srvx/commit/ef71acd))
+
+### ❤️ Contributors
+
+- Pooya Parsa ([@pi0](https://github.com/pi0))
+
 ## v0.9.3
 
 [compare changes](https://github.com/h3js/srvx/compare/v0.9.2...v0.9.3)